pyproject.toml
[project]
name = "drugs-api"
version = "0.1.0"
description = "Drugs prices scraper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "oauth2client>=4.1.3",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "google-api-core==2.24.2",
    "google-api-python-client==2.169.0",
    "google-auth==2.40.1",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "oauthlib==3.2.2",
    "proto-plus==1.26.1",
    "protobuf==6.30.2",
    "requests-oauthlib==2.0.0",
    "uritemplate==4.1.1"
]
